🔨 Refactor: userId를 전역으로 관리하도록 수정 및 동일 위치라고 판단할 경우 관제 센터 메일 API 추가

+++ WAYVI/WAYVI Watch App/Views/HomeNavigationView.swift
@@ -58,7 +58,7 @@ struct HomeNavigationView: View {
                 }
             }
             .sheet(isPresented: $fallManager.fallDetected) {
-                HealthAlertView(userId: loginViewModel.userId)
+                HealthAlertView(userId: Int64(loginViewModel.userId))
                     .environmentObject(fallManager)
             }
         }
